feat: Add `CameraDirection` to `scanBarcode`
Update `OSBarcodeLib.xcframework` to the latest version that includes the Camera Selection.
Extend the library's `OSBARCCameraModel` so that the parameter received can be mapped into the value the library expects.
Add the `cameraOrientation` property to the `OSBarcodeScanArgumentsModel` structure and use it on the library's `scanBarcode` method call.
Update the old version's Javascript method so that `cameraDirection` is transformed from the old `Text` version to the `Number` expected by the Wrapper.

References: https://outsystemsrd.atlassian.net/browse/RMET-2754