feat: Add `ScanOrientation` to `ScanBarcode` on iOS
Update the `OSBarcodeLib.xcframewor` to the latest version that includes the `scanOrientation` parameter. This is also added to the `ScanBarcode` method parameter list so that this can be fetched from the OutSystems side and passed to the iOS library.
Move the `map(value:)` method to a standalone protocol since it's used by multiple Models.
Add the `String` to `Number` conversion for the old Barcode plugin version bridge method.

References: https://outsystemsrd.atlassian.net/browse/RMET-2753